Use uint32_t for indices.
This allows us to save some memory and opens up more possibilities
for future optimizations. And 4,294,967,295 should still be more
than enough space for our lists of traces, strings etc. pp.

+++ accumulatedtracedata.h
@@ -32,7 +32,7 @@
 template<typename Base>
 struct Index
 {
-    uint64_t index = 0;
+    uint32_t index = 0;
 
     explicit operator bool() const
     {
